refactor: move model class to model package
