+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/model/GrapheneKeyMapper.kt
@@ -1,4 +1,4 @@
-package com.graphene.writer.store.key
+package com.graphene.writer.store.key.model
 
 import com.graphene.writer.input.GrapheneMetric
 import org.elasticsearch.common.xcontent.XContentBuilder
